+++ common/src/main/java/club/iananderson/seasonhud/config/DefaultValues.java
@@ -25,6 +25,7 @@ public class DefaultValues {
     public static final boolean DEFAULT_SHOW_SUB_SEASON = false;
     public static final ShowDay DEFAULT_SHOW_DAY = ShowDay.SHOW_DAY;
     public static final boolean DEFAULT_SHOW_FERTILITY = false;
+    public static final boolean DEFAULT_FERTILITY_REPLACES_SEASON = false;
     public static final boolean DEFAULT_ENABLE_MINIMAP_INTEGRATION = true;
     public static final boolean DEFAULT_SHOW_DEFAULT_WHEN_MINIMAP_HIDDEN = false;
     public static final boolean DEFAULT_JOURNEYMAP_ABOVE_MAP = false;